back/pyproject.toml
[project]
name = "trading-platform-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "fastapi<0.116.0",
    "firebase-admin",
    "google-api-python-client",
    "httpx",
    "matplotlib",
    "numpy",
    "pandas",
    "polars",
    "prefect",
    "pydantic",
    "pytest",
    "python-socketio[asgi]",
    "pyyaml",
    "requests",
    "sortedcontainers",
    "starlette<0.41.0",
    "termcolor",
    "uvicorn",
]
